chore: format package
