+++ package.json
@@ -19,7 +19,7 @@
     "dev:docs:adapters": "turbo run dev --filter=docs",
     "email": "cd apps/dev/nextjs && pnpm email",
     "eslint": "eslint --cache .",
-    "lint": "prettier --check .",
+    "lint": "prettier --check ./packages",
     "format": "prettier --write .",
     "release": "release",
     "peek": "pnpm release --peek",
@@ -31,8 +31,8 @@
     "@balazsorban/monorepo-release": "0.3.1",
     "@types/jest": "^29.5.7",
     "@types/node": "^20.8.10",
-    "@typescript-eslint/eslint-plugin": "5.47.0",
-    "@typescript-eslint/parser": "5.47.0",
+    "@typescript-eslint/eslint-plugin": "v6.19.1",
+    "@typescript-eslint/parser": "v6.19.1",
     "eslint": "8.30.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard-with-typescript": "^21.0.1",
